Improve markup and styling on the results page
The initial markup was to enable the developers to confirm the whole journey worked and looped through the data too many times. I've amended the markup to loop through the data once.

The styling is in place to provide a better demo on 11th July; it will be amended to use mas standards and BEM.