Make your contribution stack more concise
Before there was a delegator class doing the role as middle person to
convert full or minimum contribution to pass to the calculator
and then the calculator returned a third object: "your contribution"
object.

I changed to: makes Your contribution class knows
about the two converter class and return itself. IMHO,
it makes the code more precise, concise and easier to read.

Another change was the session store for eligible salary,
I prefer to wrap in a controller filter since it is not
part of the action itself and just something
to be handle on other steps.