Refactor CalculatorDelegator removing unecessary steps
Since the output of this class was the eligible salary and the
employee and employer percent, I renamed this class to
be YourContributionGenerator.

I removed the calculate method from the minimum and full contribution
and I removed YourContribution class since I think it is
an unecessary step because:

Before:

1) it were delegating to a class
2) The calculator class had a method called calculate
3) this class should have a calculate method that returns a third object.
4) This third object has exactly the same name as the attributes with
the method names of the calculators.

After:

1) Have a class to returns a contribution calculator and that's it.

Since the contribution calculators has a interface, if we respect
this interface we don't need the calculate neither a third object.