Run the tests in Jenkins under Docker
In order to run the tests for this gem in Jenkins, we need to provide
a Docker image. It runs Ubuntu 17.04, so the version of ruby is upgraded
to match the default in that OS.

In addition the gems dependencies have changed: testing needs a new gem
`tzinfo-data` in order to work; `rubocop` becomes a global dependency;
`bowndler` gets installed from the gems server (as there are no new
modifications on github); `better_errors` and `binding_of_caller` are
included in the test group.