Fix messages based on the displayed results table
The figures were wrong, the following is accurate:

Scenario 1
When an Employee and an Employer, both contributing more than the minimum
increase for 2018 AND 2019
Employee contribution >= 5%
Employer contribution >= 3%

Scenario 2
When an Employee and an Employer, both contributing more than the minimum
increase for 2018 but NOT for 2019
Employee contribution >= 3% and <5%
Employer contribution >= 2% and <3%

Scenario 3
When only an Employee contributing more than the minimum increase for 2018 AND
2019
Employee contribution >= 5%
Employer contribution = minimum % as set by the Govt. over that period

Scenario 4
When only an Employer contributing more than the minimum increase for 2018 AND
2019
Employee contribution = minimum % as set by the Govt. over that period
Employer contribution >= 3%