Convert Javascript opt in salary thresholds to use a yml file.
The Dough convention for loading components, e.g
data-dough-component=SalaryConditions, is:
the html attribute called 'data-dough' PLUS
the name of the component PLUS
'config' (e.g data-dough-salary-conditions-config).

Dough then parses the object and stores the data in
the config variable, which is the second
argument of the initialise function.

Because the javascript and backend format the salary_threshold
data differently,
it is necessary to convert the backend format to the format
required by the js.

Prior to this change, the JS 'optInTriggers' variable was on local
scope of the salary conditions file. Now it is an attribute of
the object salary conditions which necessitates using 'this' to ensure
the correct scope.