Update figures for the new tax year
Various figures are used for many of the calculations on the site to determine
eligibility for auto-enrollment and how much tax relief is due. These figures
are also used to show various messages to aid user understanding of auto-enrolment.

Consequently, alot of tests needed amending. I also removed some steps for future
periods which are now redundant.